--- src/components/mod.rs
@@ -1,3 +1,0 @@
-pub mod conversation_area;
-pub mod prompt_area;
-pub mod sidebar;
