--- src/components/sidebar.rs
@@ -1,6 +1,0 @@
-use leptos::{component, view, IntoView};
-
-#[component]
-pub fn Sidebar() -> impl IntoView {
-    view! {}
-}
